+++ vue-project/src/views/admin/AdminClientsAssignPharmaciesView.vue
@@ -46,7 +46,7 @@
           <button class="btn-excell-template" @click="downloadTemplate" style="margin-right: 1rem;">엑셀 템플릿</button>
                     <button class="btn-excell-upload" @click="triggerFileUpload" style="margin-right: 1rem;">엑셀 등록</button>
           <button class="btn-excell-download" @click="downloadExcel" style="margin-right: 1rem;">엑셀 다운로드</button>
-          <button class="btn-delete" @click="deleteAllAssignments">모두 삭제</button>
+<!--          <button class="btn-delete" @click="deleteAllAssignments">모두 삭제</button>-->
           <input
             ref="fileInput"
             type="file"
@@ -465,6 +465,10 @@ async function assignPharmacies() {
   await fetchClients()
 }
 async function deleteAssignment(client, pharmacy = null) {
+  if (!confirm('정말 삭제하시겠습니까?')) {
+    return
+  }
+
   let query = supabase.from('client_pharmacy_assignments').delete().eq('client_id', client.id)
   if (pharmacy) query = query.eq('pharmacy_id', pharmacy.id)
   await query
@@ -783,16 +787,16 @@ const downloadExcel = async () => {
 
 
 
-async function deleteAllAssignments() {
-  if (!confirm('정말 모든 문전약국 지정 데이터를 삭제하시겠습니까?\n이 작업은 되돌릴 수 없습니다.')) return;
-  const { error } = await supabase.from('client_pharmacy_assignments').delete().neq('id', 0);
-  if (error) {
-    alert('삭제 중 오류가 발생했습니다: ' + error.message);
-    return;
-  }
-  clients.value.forEach(c => c.pharmacies = []);
-  alert('모든 문전약국 지정 데이터가 삭제되었습니다.');
-}
+// async function deleteAllAssignments() {
+//   if (!confirm('정말 모든 문전약국 지정 데이터를 삭제하시겠습니까?\n이 작업은 되돌릴 수 없습니다.')) return;
+//   const { error } = await supabase.from('client_pharmacy_assignments').delete().neq('id', 0);
+//   if (error) {
+//     alert('삭제 중 오류가 발생했습니다: ' + error.message);
+//     return;
+//   }
+//   clients.value.forEach(c => c.pharmacies = []);
+//   alert('모든 문전약국 지정 데이터가 삭제되었습니다.');
+// }
 
 // 오버플로우 감지 및 툴팁 제어 함수들
 const checkOverflow = (event) => {
